# pkg/plugins: Guard storage and unparallel racy tests